勤務時間表示の数値がおかしい件、データベースのwork_reportsテーブルのwork_timesの型が間違っていた(int16 -> double)

+++ NIPPO/MonthlyReport.cs
@@ -235,15 +235,15 @@ namespace NIPPO
             this._monthDs.Tables.Add(_dt);
             return this._monthDs;
         }
-        public int getTotalTimeCommon(string columnName)
+        public double getTotalTimeCommon(string columnName)
         {
-            int sum = 0;
+            double sum = 0.0;
             for (int i = 0; i < this._monthDs.Tables[this.listTableName].Rows.Count; i++)
             {
                 object workTimes = this._monthDs.Tables[this.listTableName].Rows[i][columnName];
                 if (workTimes != null && workTimes.ToString() != "")
                 {
-                    sum += int.Parse(workTimes.ToString());
+                    sum += double.Parse(workTimes.ToString());
                 }
             }
             return sum;
@@ -254,7 +254,7 @@ namespace NIPPO
             return(this.getTotalTime().ToString() + " h");
         }
 
-        public int getTotalTime()
+        public double getTotalTime()
         {
             return(this.getTotalTimeCommon("work_times"));
         }
@@ -264,7 +264,7 @@ namespace NIPPO
             return (this.getTotalOverTime125().ToString() + " h");
         }
 
-        public int getTotalOverTime125()
+        public double getTotalOverTime125()
         {
             return (this.getTotalTimeCommon("overtime125"));
         }
@@ -274,7 +274,7 @@ namespace NIPPO
             return (this.getTotalOverTime150().ToString() + " h");
         }
 
-        public int getTotalOverTime150()
+        public double getTotalOverTime150()
         {
             return (this.getTotalTimeCommon("overtime150"));
         }
